ESECLDTPLT-3119 - Standardize resource names

+++ automated-test-scripts/f5-aws-cloudformation-v2/examples/quickstart/common/verify_waf_blocking.sh
@@ -3,13 +3,18 @@
 #  replayEnabled = true
 #  replayTimeout = 10
 
-dnsAppname=$(aws cloudformation describe-stacks --stack-name <STACK NAME> --region <REGION> | jq -r '.Stacks[].Outputs[] | select (.OutputKey=="vipPublicUrl") | .OutputValue')
 
-echo "Attempting illegal action (enforcement mode should be set to blocking by default) on ${dnsAppname}"
-
-REJECTED_RESPONSE=$(curl -sk -X DELETE ${dnsAppname})
-echo "REJECTED_RESPONSE: ${REJECTED_RESPONSE}"
-
-if echo ${REJECTED_RESPONSE} | grep -q "The requested URL was rejected"; then
+if [[ "<PROVISION EXAMPLE APP>" == 'false' ]]; then
     echo "PASS"
+else
+    dnsAppname=$(aws cloudformation describe-stacks --stack-name <STACK NAME> --region <REGION> | jq -r '.Stacks[].Outputs[] | select (.OutputKey=="vipPublicUrl") | .OutputValue')
+
+    echo "Attempting illegal action (enforcement mode should be set to blocking by default) on ${dnsAppname}"
+
+    REJECTED_RESPONSE=$(curl -sk -X DELETE ${dnsAppname})
+    echo "REJECTED_RESPONSE: ${REJECTED_RESPONSE}"
+
+    if echo ${REJECTED_RESPONSE} | grep -q "The requested URL was rejected"; then
+        echo "PASS"
+    fi
 fi
